refactor: Split MenuBarView.swift into separate files
