+++ InputMetrics/InputMetrics/Views/MetricTab.swift
@@ -1,0 +1,5 @@
+import SwiftUI
+
+enum MetricTab {
+    case mouse, keyboard
+}
